+++ crates/ursa-gateway/src/admin/route/api/v1/post.rs
@@ -4,7 +4,7 @@ use axum::Extension;
 use hyper::StatusCode;
 use tokio::sync::RwLock;
 
-use crate::cache::tlrfu::Tlrfu;
+use crate::cache::Tlrfu;
 
 pub async fn purge_cache_handler(Extension(cache): Extension<Arc<RwLock<Tlrfu>>>) -> StatusCode {
     cache.write().await.purge();
