+++ crates/ursa-gateway/src/cache/mod.rs
@@ -1,2 +1,3 @@
 mod lru;
-pub mod tlrfu;
+mod tlrfu;
+pub use tlrfu::*;
