+++ crates/ursa-gateway/src/main.rs
@@ -9,7 +9,7 @@ mod worker;
 use std::{path::PathBuf, str::FromStr, sync::Arc};
 
 use anyhow::{Context, Result};
-use cache::tlrfu::Tlrfu;
+use cache::Tlrfu;
 use clap::Parser;
 use cli::{Cli, Commands};
 use tokio::{sync::RwLock, task};
@@ -43,7 +43,10 @@ async fn main() -> Result<()> {
             // sync
             gateway_config.merge_daemon_opts(opts);
 
-            let server_cache = Arc::new(RwLock::new(Tlrfu::new(gateway_config.cache.max_size, 0)));
+            let server_cache = Arc::new(RwLock::new(Tlrfu::new(
+                gateway_config.cache.max_size,
+                gateway_config.cache.ttl_buf as u128 * 1_000_000, // ms to ns
+            )));
             let admin_cache = server_cache.clone();
 
             let server_config = Arc::new(RwLock::new(gateway_config));
